# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AITT Compliance Backend API
  version: 1.0.0
  description: >
    REST API for the AITT (AI Transparency Token) compliance-certification
    platform. Documents are submitted by companies, reviewed by sub-admins
    (0–100 compliance score + one of four review decisions), certified by the
    Main Admin, and anchored on the deployed Stellar/Soroban contract
    (CA6KYPPXEUTAP4X6JAEOI37OD2SCKEAUOSV2VN5ICDWCAI4WASFHRSYB). Exceptional
    actions (revocation, threshold changes, upgrades) go through N-of-M
    multi-sig governance proposals.

    Auth: obtain a JWT via POST /auth/login, then send it as
    `Authorization: Bearer <accessToken>` on protected routes.

    Health check (outside /api/v1): `GET /` returns a plain-text running message.
servers:
  - url: https://aitt-backend.vercel.app/api/v1
    description: Vercel (testnet)
  - url: http://localhost:4000/api/v1
    description: Local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Forbidden:
      description: Authenticated but role not permitted
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        details: { type: object, nullable: true }
...
security:
  - bearerAuth: []
paths:
...
  /certificates:
    post:
      tags: [Certificates]
      summary: Create a certificate (super_admin, regulator_admin)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [certificateName, subject]
              properties:
                certificateName: { type: string }
                subject: { type: string }
                file: { type: string, format: binary, nullable: true }
      responses:
        '201': { description: Certificate created }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /certificates/{id}:
    get:
      tags: [Certificates]
      summary: Get a certificate (optional auth adds private fields)
      security: []
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Certificate }
        '404': { $ref: '#/components/responses/NotFound' }
...